+++ src/btviewer/route.py
@@ -1,0 +1,5 @@
+from btviewer import route
+
+@app.route('/')
+def index():
+    return 'Hello World!'
